RTMsim

RTMsim is a new software tool for RTM filling simulations which fulfills these requirements: Several test cases were used for successfully validating the implemented model. The porous cavity is fully described by a mesh file with triangular cells on the part’s mid-surface and cell set definitions. The latter can be used for specifying the location of the pressure injection ports and regions with different preforms by assigning different thickness, porosity and permeability values. Additional equations (e.g. for modeling the degree-of-cure) can either be added with equations of the same type or modifications of existing equations (e.g. for variable cavity thickness as needed for vacuum assisted resin infusion simulations).

Package Features

The simulation model shall give correct results for filling pattern and filling time.

The simulation tool takes only composite-manufacturing related inputs and the simulation shall be robust independent of numerics-related input.

The simulation tool takes a shell model of the geometry as input and the location-dependent properties are assigned directly on the shell elements.

New functionalities can be implemented by either adding equations of the same type or modifying existing equations.

Function Documentation

rtmsim.plot_mesh — Method

function plot_mesh(meshfilename,i_mode)

Create mesh plot with cells with i_mode==1 and create mesh plots with cell center nodes with i_mode==2 for manual selection of inlet ports.

Unit test:

rtmsim.plot_mesh("..\meshfiles\mesh_permeameter1_foursets.bdf",1)

Additional unit tests:

rtmsim.plot_mesh("..\meshfiles\mesh_annulusfiller1.bdf",2) for the manual selection of inlet ports with left mouse button click while key p is pressed

rtmsim.rtmsim_rev1(1,"..\meshfiles\mesh_annulusfiller1.bdf",200, 0.35e5,1.205,1.4,0.06, 0.35e5,0.00e5, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-10,1,1,0,0, 0,0,0,0, 0,"results.jld2",1,0.01,16) for starting only with the interactively selected inlet ports

rtmsim.read_nastran_mesh — Method

function read_nastran_mesh(meshfilename,paramset,paramset1,paramset2,paramset3,paramset4,patchtype1val,patchtype2val,patchtype3val,patchtype4val,i_interactive,r_p)

Read file in Nastran format with fixed length (8 digits), nodes (GRIDS) defined in global coordinate system.

Unit test:

paramset=[0.5,0.3,3e-10,1.0,1.0,0.0,0.0];paramset1=paramset;paramset2=paramset;paramset3=paramset;paramset4=paramset;patchtype1val=-1;patchtype2val=-1;patchtype3val=-1;patchtype4val=-1;i_interactive=0;r_p=0.01; N,cellgridid,gridx,gridy,gridz,cellcenterx,cellcentery,cellcenterz,patchparameters,patchparameters1,patchparameters2,patchparameters3,patchparameters4,patchids1,patchids2,patchids3,patchids4,inletpatchids=read_mesh(meshfilename,paramset,paramset1,paramset2,paramset3,paramset4,patchtype1val,patchtype2val,patchtype3val,patchtype4val,i_interactive,r_p)`

rtmsim.start_rtmsim — Method

start_rtmsim(inputfilename)

Reads the text input file and calls the solver with the read parameters. The complete set of input parameters can be accessed in the input file. The following paragraph shows an example for such an input file:

    1    #i_model 
    meshfiles\mesh_permeameter1_foursets.bdf    #meshfilename 
    200    #tmax 
    1.01325e5 1.225 1.4 0.06    #p_ref rho_ref gamma mu_resin_val 
    1.35e5 1.0e5    #p_a_val p_init_val 
    3e-3 0.7 3e-10 1 1 0 0    #t_val porosity_val K_val alpha_val refdir1_val refdir2_val refdir3_val 
    3e-3 0.7 3e-10 1 1 0 0    #t1_val porosity1_val K1_val alpha1_val refdir11_val refdir21_val refdir31_val 
    3e-3 0.7 3e-10 1 1 0 0    #t2_val porosity2_val K2_val alpha2_val refdir12_val refdir22_val refdir32_val
    3e-3 0.7 3e-10 1 1 0 0    #t3_val porosity3_val K3_val alpha3_val refdir13_val refdir23_val refdir33_val
    3e-3 0.7 3e-10 1 1 0 0    #t4_val porosity4_val K4_val alpha4_val refdir14_val refdir24_val refdir34_val 
    1 0 0 0    #patchtype1val patchtype2val patchtype3val patchtype4val 
    0 results.jld2    #i_restart restartfilename
    0 0.01    #i_interactive r_p
    16    #n_pics

Meaning of the variables:

i_model: Identifier for physical model (Default value is 1)

meshfilename: Mesh filename.

tmax: Maximum simulation time.

p_ref rho_ref gamma mu_resin_val: Parameters for the equation of state and dynamic viscosity of resin used in the Darcy term.

p_a_val p_init_val: Absolut pressure value for injection port and for initial cavity pressure.

t_val porosity_val K_val alpha_val refdir1_val refdir2_val refdir3_val: Properties of the cells in the main preform: The vector (refdir1_val,refdir2_val,refdir3_val) is projected onto the cell in order to define the first principal cell direction. The second principal cell direction is perpendicular to the first one in the plane spanned by the cell nodes. The principal cell directions are used as the principal permeabilty directions. The cell properties are defined by the thickness t_val, the porosity porosity_val, the permeability K_val in the first principal cell direction, the permeablity alpha_val in the second principal direction.

t1_val porosity1_val K1_val alpha1_val refdir11_val refdir21_val refdir31_val etc.: Properties for up to four additional cell regions if preform.

patchtype1val patchtype2val patchtype3val patchtype4val: These regions are used to specify the location of the pressure boundary conditions and to specify regions with different permeability, porosity and thickness properties (e.g. for different part thickness and layup or for race tracking which are regions with very high permeability typically at the boundary of the preforms). Vents need not be specified. Parameters patchtype1val define the patch type.

i_restart restartfilename: Start with new simulation if 0 or continue previous simulation if 1.

i_interactive r_p: Select the inlet ports graphically if i_interactive equal to 1.

n_pics: Number of intermediate output files. Supposed to be a multiple of 4.

Entries are separated by one blank.

Unit test:

rtmsim.start_rtmsim("..\inputfiles\input.txt")

rtmsim.rtmsim_rev1 — Method

rtmsim_rev1(i_model,meshfilename,tmax,
            p_ref,rho_ref,gamma,mu_resin_val,
            p_a_val,p_init_val,
            t_val,porosity_val,K_val,alpha_val,refdir1_val,refdir2_val,refdir3_val,
            t1_val,porosity1_val,K1_val,alpha1_val,refdir11_val,refdir21_val,refdir31_val,
            t2_val,porosity2_val,K2_val,alpha2_val,refdir12_val,refdir22_val,refdir32_val,
            t3_val,porosity3_val,K3_val,alpha3_val,refdir13_val,refdir23_val,refdir33_val,
            t4_val,porosity4_val,K4_val,alpha4_val,refdir14_val,refdir24_val,refdir34_val,
            patchtype1val,patchtype2val,patchtype3val,patchtype4val,i_restart,restartfilename,i_interactive,r_p,n_pics)

RTMsim solver with main steps

Simulation initialization

Read mesh file and prepare patches

Find neighbouring cells

Assign parameters to cells

Create local cell coordinate systems

Calculate initial time step

Array initialization

Define simulation time and intermediate output times

Boundary conditions

(Optional initialization if i_model=2,3,..)

Time evolution (for loops over all indices inside a while loop for time evolution)

Calculation of correction factors for cell thickness, porosity, permeability, viscosity

Pressure gradient calculation

Numerical flux function calculation

Update of rho, u, v, gamma and p according to conservation laws and equation of state

Boundary conditions

Prepare arrays for next time step

Saving of intermediate data

(Opional time marching etc. for i_model=2,3,...)

Calculation of adaptive time step

Unit tests:

rtmsim.rtmsim_rev1(1,"..\meshfiles\mesh_permeameter1_foursets.bdf",200, 101325,1.225,1.4,0.06, 1.35e5,1.00e5, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-11,1,1,0,0, 3e-3,0.7,3e-11,1,1,0,0, 3e-3,0.7,3e-9,1,1,0,0, 1,2,2,2,0,"results.jld2",0,0.01,16)

Addtional unit tests:

rtmsim.rtmsim_rev1(1,"..\meshfiles\mesh_permeameter1_foursets.bdf",200, 101325,1.225,1.4,0.06, 1.35e5,1.00e5, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-11,1,1,0,0, 3e-3,0.7,3e-11,1,1,0,0, 3e-3,0.7,3e-9,1,1,0,0, 1,0,0,0, 0,"results.jld2",0,0.01,16) for starting a simulation with one pressure inlet port (sets 2, 3 and 4 are not used and consequently the preform parameters are ignored; since set 1 is a pressure inlet, also the parameters for set 1 are ignored and the only relevant parameter for the specified set is the pressure difference between injection and initial cavity pressure)

rtmsim.rtmsim_rev1(1,"..\meshfiles\mesh_permeameter1_foursets.bdf",200, 101325,1.225,1.4,0.06, 1.35e5,1.00e5, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-11,1,1,0,0, 3e-3,0.7,3e-11,1,1,0,0, 3e-3,0.7,3e-9,1,1,0,0, 1,2,2,2, 0,"results.jld2",0,0.01,16) for starting a simulation with different patches and race tracking

rtmsim.rtmsim_rev1(1,"..\meshfiles\mesh_permeameter1_foursets.bdf",200, 101325,1.225,1.4,0.06, 1.35e5,1.00e5, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-11,1,1,0,0, 3e-3,0.7,3e-11,1,1,0,0, 3e-3,0.7,3e-9,1,1,0,0, 1,2,2,2, 1,"results.jld2",0,0.01,16) for continuing the previous simulation

rtmsim.plot_results — Method

function plot_results(resultsfilename)

Create contour plots of the filling factor and the pressure after loading a results file.

Unit test:

rtmsim.plot_results("../results.jld2")

rtmsim.plot_overview — Method

function plot_overview(n_out,n_pics)

Create filling factor contour plots. n_out is the index of the last output file, if n_out==-1 the output file with the highest index is chosen. Consider the last n_pics for creating the contour plots at four equidistant time intervals, if n_pics==-1 all available output files are considered.

Unit test:

rtmsim.plot_overview(-1,-1)

rtmsim.plot_filling — Method

function plot_filling(n_out,n_pics)

Create a window showing the filling factor contour plot at a selected time instance. Selection is with slider bar. n_out is the index of the last output file, if n_out==-1 the output file with the highest index is chosen. Consider the last n_pics for creating the contour plots at four equidistant time intervals, if n_pics==-1 all available output files are considered.

Unit test:

rtmsim.plot_filling(-1,-1)

rtmsim.gui — Method

function gui()

Opens the GUI.